requirements.txt
appdirs==1.4.4
astroid==2.4.2
audioread==2.1.8
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
decorator==4.4.2
idna==2.10
isort==5.5.4
joblib==0.17.0
lazy-object-proxy==1.4.3
librosa==0.8.0
llvmlite==0.34.0
mccabe==0.6.1
numba==0.51.2
numpy==1.19.2
opencv-python==4.4.0.42
packaging==20.4
Pillow==8.0.1
pooch==1.2.0
pycairo==1.20.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
requests==2.24.0
resampy==0.2.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
